(#208) refactor: mock 데이터 분리
